Move option-defaulting to option-manager.
